prosim, prosim, mergni sa 🙏🙏🙏

+++ Platforms.cs
@@ -6,11 +6,11 @@ namespace CiganSimulator
     {
         public int x;
         public int y;
-        public int width;
-        public int height;
+        public float width;
+        public float height;
         private Vector2 mapVelocity;
 
-        public Platform(int x, int y, int width, int height)
+        public Platform(int x, int y, float width, float height)
         {
             this.x = x;
             this.y = y;
